+++ src/mod.cpp
@@ -154,10 +154,26 @@ void LoadGameSettings(const IniFile* settingsIni)
     const bool eggCarrierTransformationCutscene = settingsIni->getBool("GameSettings", "EggCarrierTransformationCutscene", true);
     const bool skipCredits = settingsIni->getBool("GameSettings", "SkippableCredits", true);
     const bool winButtonEnabled = settingsIni->getBool("GameSettings", "AutoWinButton", false);
+    
+    const bool eggmanCommentOnTrap = settingsIni->getBool("CharacterVoiceReactions", "EggmanOnTrap", false);
+    const bool otherCharactersCommentOnTrap = settingsIni->getBool("CharacterVoiceReactions", "OtherCharactersOnTrap", false);
+    const bool currentCharacterReactToTrap = settingsIni->getBool("CharacterVoiceReactions", "CurrentCharacterOnTrap", false);
+    
+    const bool eggmanCommentOnCharacterUnlock = settingsIni->getBool("CharacterVoiceReactions", "EggmanOnUnlock", false);
+    const bool currentCharacterCommentOnCharacterUnlock = settingsIni->getBool("CharacterVoiceReactions", "CurrentCharacterOnUnlock", false);
+    const bool unlockedCharacterCommentOnCharacterUnlock = settingsIni->getBool("CharacterVoiceReactions", "UnlockedCharactersOnUnlock", false);
+    const bool eggmanCommentOnKeyItems = settingsIni->getBool("CharacterVoiceReactions", "EggmanOnKeyItem", false);
+    const bool tikalCommentOnKeyItems = settingsIni->getBool("CharacterVoiceReactions", "TikalOnKeyItem", false);
+    const bool currentCharacterCommentOnKeyItems = settingsIni->getBool("CharacterVoiceReactions", "CurrentCharacterOnKeyItem", false);
 
     cheatsManager.SetCheatsConfiguration(autoSkipCutscenes, skipCredits, winButtonEnabled);
     eventDetector.SetMultipleMissions(completeMultipleLevelMissions);
     worldStateManager.SetEggCarrierTransformationCutscene(eggCarrierTransformationCutscene);
+    characterManager.SetCharacterVoiceReactions(eggmanCommentOnTrap, otherCharactersCommentOnTrap, currentCharacterReactToTrap);
+
+    randomizer.SetCharacterVoiceReactions(eggmanCommentOnCharacterUnlock, currentCharacterCommentOnCharacterUnlock,
+                                          unlockedCharacterCommentOnCharacterUnlock, eggmanCommentOnKeyItems,
+                                          tikalCommentOnKeyItems, currentCharacterCommentOnKeyItems);
 }
 
 #define ReplacePNG_Common(a) do { \