+++ CHANGELOG.md
@@ -8,7 +8,9 @@ All notable changes to this project will be documented in this file.
 
 - Created [Location Wiki](https://github.com/ClassicSpeed/SADX-Archipelago-Locations-Wiki/blob/main/README.md) for the
   locations of the life capsules and field emblems. (Credit to DementisXYZ).
-- Added random voice lines when getting a trap.
+- Added voice lines when getting a trap, a key item or a new character.
+    - The voice lines are from Eggman or the current character.
+    - There are settings in the mod configuration to disable them.
 - Re-enabled the Egg Carrier transformation cutscene to help players know that the switch worked.
     - The cutscene is unskippable, so an option in the mod configuration was added to disable it.
 - When starting on Angel Island, the player will spawn near the Red Mountain entrance.